Added some stuff
I found out that I can install multiple programs with a single command just by putting them on the same line, so I changed the code to just store all the applications in a list by name. I'll grab the names from there and just yeet them all in somehow.

I also realized I may use a different package manager. So I made a note of that for later.

The print statement is just a placeholder for now.